// src/app/Interfaces/entretien-source-eau.interface.ts
export interface EntretienSourceEauInterface {
  idEntretienSourceEau?: string;
  nomEntretienSourceEau: string;
  descriptionEntretienSourceEau: string;
  chargeEntretienSourceEau: number;
  isExternalEntretienSourceEau: boolean;
  descriminant?: string;
  nomSourceEau: string;
  nomFournisseur: string;
}